% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_JSD_subset_fun.R
\name{calculate_JSD_subset}
\alias{calculate_JSD_subset}
\title{This function takes in 2 matrices of equal columns and calculates all pairwise JSD between them}
\usage{
calculate_JSD_subset(prediction, syn_spots_profiles, JSD_indices)
}
\arguments{
\item{prediction:}{Object of class matrix with n ((n spots)) rows and k (n topics) columns where each column is a probabilistic array}

\item{syn_spots_profiles:}{Object of class matrix with m (n synthetic spots) rows and k (n topics) columns where each column is a probabilistic array}

\item{JSD_indices:}{Object of class list of m (n spots) elements with with each element having row I indices of syn_spots_profiles}
}
\value{
this function returns a matrix with n rows and I columns the JSD values for the comparisons performed for each prediction
}
\description{
This function takes in 2 matrices of equal columns and calculates all pairwise JSD between them
}
\examples{

}
