% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_markers2_fun.R
\name{cut_markers2}
\alias{cut_markers2}
\title{Select the top n cluster markers by ranking the genes by pvalue and then by avglogFC}
\usage{
cut_markers2(markers, ntop)
}
\arguments{
\item{markers}{A data.frame as in the output of the FindAllMarkers Seurat function.}

\item{ntop}{The number of top markers you want.}
}
\value{
A dataframe with the ntop markers per cluster with the zscore of the logFC.
}
\description{
Select the top n cluster markers by ranking the genes by pvalue and then by avglogFC
}
\examples{

}
