% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatial_interaction_graph_fun.R
\name{get_spatial_interaction_graph}
\alias{get_spatial_interaction_graph}
\title{This function takes in deconvolution matrix and returns a network igraph object}
\usage{
get_spatial_interaction_graph(decon_mtrx)
}
\arguments{
\item{decon_mtrx}{Object of class matrix, SPOTxCELL-TYPE, with the cell-type proportions per spot.}
}
\description{
This function takes in deconvolution matrix and returns a network igraph object
}
\examples{

}
