% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_prediction_fun.R
\name{lda_prediction}
\alias{lda_prediction}
\title{This function processes the data and trains the LDA model}
\usage{
lda_prediction(lda_mod, spot_counts, ncores)
}
\arguments{
\item{lda_mod}{Object of class LDA_Gibbs.}

\item{spot_counts}{Object of class sparse matrix. The count matrix of the spots we want to predict as SPOTxGENES}

\item{ncores}{Object of class integer, number of cores to use when parallelizing, by default it is 60\% of the detected cores.}
}
\value{
matrix with the topic probability distribution for each spot.
}
\description{
This function processes the data and trains the LDA model
}
\examples{

}
