% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_deconvolution_nmf_fun.R
\name{mixture_deconvolution_nmf}
\alias{mixture_deconvolution_nmf}
\title{Run mixtures through the NMF model to get the cell type composition.}
\usage{
mixture_deconvolution_nmf(
  nmf_mod,
  mixture_transcriptome,
  transf,
  reference_profiles,
  min_cont = 0.01
)
}
\arguments{
\item{nmf_mod}{Object of class NMF containing the trained nNMF model.}

\item{mixture_transcriptome}{Object of class matric of dimensions GENESxSPOTS}

\item{transf}{Transformation to normalize the count matrix: uv (unit variance), raw (no transformation applied). By default UV.}

\item{reference_profiles}{Object of class matrix containing the TOPICSxCELLS Coefficient matrix from where want to get the weights. It can be cell type profiles or cell specific profiles.}

\item{min_cont}{Object of class numeric; Indicates the minimum contribution we expect from a cell in that spot. Since we're working with proportions by setting 0.01, by default, means that we will accept those cell types whose weight coefficient is at least 1\% of the total.}
}
\value{
This function returns a matrix with the coefficients of the spatial mixtures.
}
\description{
Run mixtures through the NMF model to get the cell type composition.
}
\examples{

}
