% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_image_fun.R
\name{plot_image}
\alias{plot_image}
\title{This function takes in a seurat object, a path to an image and a slice name to load and plot.}
\usage{
plot_image(img_path, img_alpha = 1)
}
\arguments{
\item{img_path:}{Object of class string pointing to the HE image.}

\item{img_alpha:}{Object of class numeric between 0-1 indicating the degree of transparency of the image}
}
\value{
this function returns a plot of class gg
}
\description{
This function takes in a seurat object, a path to an image and a slice name to load and plot.
}
\examples{

}
