% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotImage.R
\name{test}
\alias{test}
\alias{plotImage,character-method}
\alias{plotImage,array-method}
\title{Plot JP(E)G/PNG image}
\usage{
\S4method{plotImage}{character}(x)

\S4method{plotImage}{array}(x)
}
\arguments{
\item{x}{character string corresponding to an image file path.
Valid file types are JPG, JPEG and PNG.}

\item{alpha}{single numeric in between 0 and 1 determining the
image opacity. Lower values correspond to more transparency.}
}
\value{
\code{ggplot} object
}
\description{
...
}
\examples{
plotImage("~/packages/SPOTlight/inst/extdata/SPOTlight.png")
}
