% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScatterPie.R
\name{plotScatterPie,matrix-method}
\alias{plotScatterPie,matrix-method}
\title{Plot JP(E)G/PNG image}
\usage{
\S4method{plotScatterPie}{matrix}(
  x,
  groups = NULL,
  highlight = NULL,
  image = NULL,
  alpha = 1,
  scale = 0.5
)
}
\value{
\code{ggplot} object
}
\description{
...
}
\examples{
img <- "~/Packages/SPOTlight/inst/spatial/tissue_lowres_image.png"

}
