% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpatialScatterpie.R
\name{plotSpatialScatterpie}
\alias{plotSpatialScatterpie}
\alias{plotSpatialScatterpie,matrix,matrix-method}
\alias{plotSpatialScatterpie,Seurat,ANY-method}
\alias{plotSpatialScatterpie,SpatialExperiment,ANY-method}
\alias{plotSpatialScatterpie,data.frame,ANY-method}
\alias{plotSpatialScatterpie,ANY,data.frame-method}
\alias{plotSpatialScatterpie,ANY,ANY-method}
\title{Spatial scatterpie}
\usage{
\S4method{plotSpatialScatterpie}{matrix,matrix}(
  x,
  y,
  ...,
  cell_types = colnames(y),
  img = FALSE,
  scatterpie_alpha = 1,
  pie_scale = 0.4
)

\S4method{plotSpatialScatterpie}{Seurat,ANY}(x, y, ..., slice = Images(x)[1], img = FALSE)

\S4method{plotSpatialScatterpie}{SpatialExperiment,ANY}(
  x,
  y,
  ...,
  slice = imgData(x)[1, "sample_id"],
  img = FALSE
)

\S4method{plotSpatialScatterpie}{data.frame,ANY}(x, y, ...)

\S4method{plotSpatialScatterpie}{ANY,data.frame}(x, y, ...)

\S4method{plotSpatialScatterpie}{ANY,ANY}(x, y, ...)
}
\arguments{
\item{x}{Object containig the spots coordinates, it can be an object of class
SpatialExperiment, Seurat, dataframe or matrix. For the latter two
rownames should have the spot barcodes to match x.}

\item{y}{Matrix or dataframe containing the deconvoluted spots. rownames
hould have the spot barcodes to match x.}

\item{...}{additional parameters to geom_scatterpie}

\item{cell_types}{Vector of cell type names to plot. By default uses the
column names of y.}

\item{img}{Logical TRUE or FALSE indicating whether to plot the image or not.
Objects of classes accepted by \code{plotImage} can also be passed and
that image will be used. By default FALSE.}

\item{scatterpie_alpha}{Numeric scalar to set the alpha of the pie charts.
By default 1.}

\item{pie_scale}{Numeric scalar to set the size of the pie charts.
By default 0.4.}

\item{slice}{Character string indicating which slice to plot if img is TRUE.
By default uses the first image.}
}
\value{
\code{ggplot} object
}
\description{
This function takes in the coordinates of the spots and the
  proportions of the cell types within each spot. It returns a plot where
  each spot is a piechart showing proportions of the cell type composition.
}
\examples{
set.seed(321)
# Coordinates
x <- replicate(2, rnorm(100))
rownames(x) <- paste0("spot", seq_len(nrow(x)))
colnames(x) <- c("imagecol", "imagerow")
# Proportions
y <- replicate(m <- 5, runif(nrow(x), 0, 1))
y <- prop.table(y, 1)
rownames(y) <- paste0("spot", seq_len(nrow(y)))
colnames(y) <- paste0("type", seq_len(ncol(y)))
(plt <- plotSpatialScatterpie(x = x, y = y))
}
\author{
Marc Elosua Bayes & Helena L Crowell
}
