% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopicProfiles.R
\name{plotTopicProfiles,NMF,character-method}
\alias{plotTopicProfiles,NMF,character-method}
\title{Plot NMF topic profiles}
\usage{
\S4method{plotTopicProfiles}{NMF,character}(x, y, facet = FALSE, min_prop = 0.1, ncol = NULL)
}
\arguments{
\item{x}{\code{\link{NMFfit}} object}

\item{y}{vector of group labels. Should be of length \code{ncol(coef(x))}.}

\item{facet}{logical indicating whether to stratify by group. 
If \code{FALSE}, weights will be averaged across cells for each group.}

\item{min_prop}{scalar in [0,1]. When \code{facet = TRUE}, 
only cells with a weight > \code{min_prop} will be included.}

\item{ncol}{integer scalar specifying the number of facet columns.}
}
\value{
\code{ggplot} object
}
\description{
...
}
\examples{
print("todo")

}
