% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_spatial_mixtures_nmf_fun.R
\name{predict_spatial_mixtures_nmf}
\alias{predict_spatial_mixtures_nmf}
\title{Run test spots through the basis to get the pertinent coefficients. To do this for every spot we are going to set up a system of linear equations where we need to find the coefficient, we will use non-negative least squares to determine the best coefficient fit. Returns a matrix of KxnSPOTS dimensions.}
\usage{
predict_spatial_mixtures_nmf(nmf_mod, mixture_transcriptome, transf)
}
\arguments{
\item{nmf_mod}{Object of class dataframe obtained from the function Seurat::FindAllMarkers().}

\item{mixture_transcriptome}{Object of class matric of dimensions GENESxSPOTS}

\item{transf}{Transformation to normalize the count matrix: cpm (Counts per million), uv (unit variance), raw (no transformation applied).}
}
\value{
This function returns a matrix with the coefficients of the spatial mixtures.
}
\description{
Run test spots through the basis to get the pertinent coefficients. To do this for every spot we are going to set up a system of linear equations where we need to find the coefficient, we will use non-negative least squares to determine the best coefficient fit. Returns a matrix of KxnSPOTS dimensions.
}
\examples{

}
