% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_seobj_topic_fun.R
\name{prep_seobj_topic_fun}
\alias{prep_seobj_topic_fun}
\title{This functions takes in a seurat object and returns the transposed count matrix with the highly variable genes selected}
\usage{
prep_seobj_topic_fun(se_obj)
}
\arguments{
\item{se_obj}{Object of class Seurat with the data of interest}
}
\value{
A dataframe with the ntop markers per cluster with the zscore of the logFC.
}
\description{
This functions takes in a seurat object and returns the transposed count matrix with the highly variable genes selected
}
\examples{

}
