% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterpie_plot_fun.R
\name{scatterpie_plot}
\alias{scatterpie_plot}
\title{This function takes in a seurat object and cell types of interest and returns a scatterpie plot with each spot situated in its spatial location.}
\usage{
scatterpie_plot(
  se_obj,
  cell_types_all,
  slice = NULL,
  scatterpie_alpha = 1,
  cell_types_interest = NULL,
  pie_scale = 0.4
)
}
\arguments{
\item{se_obj:}{Object of class Seurat with the spatial data and the cell type proportions in the metadata.}

\item{cell_types_all:}{Object of class vector containing the names of all the cell types.}

\item{slice:}{Object of class character, name of the slice image to load as found in se_obj@images, by default it will grab the first one on the list.}

\item{scatterpie_alpha:}{Object of class numeric between 0-1 indicating the degree of transparency of the scatterpie}

\item{cell_types_interest:}{Object of class vector containing the cell types of interest you want to plot. By setting this parameters only spots containing at least one of these cell types will be plotted. By default, NULL, it will be assumed to be the same as cell_types_all.}

\item{pie_scale:}{Object of class numeric containing the scale of the individual pie charts.}
}
\value{
this function returns a plot composition of class gg
}
\description{
This function takes in a seurat object and cell types of interest and returns a scatterpie plot with each spot situated in its spatial location.
}
\examples{

}
