% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_init_mtrx_nmf_fun.R
\name{seed_init_mtrx_nmf}
\alias{seed_init_mtrx_nmf}
\title{This functions seeds initialization matrices H and W to perform NMF}
\usage{
seed_init_mtrx_nmf(cluster_markers, se_obj, clust_vr, ntop = NULL)
}
\arguments{
\item{cluster_markers}{Object of class dataframe obtained from the function Seurat::FindAllMarkers().}

\item{se_obj}{Object of class Seurat with the data of interest.}

\item{ntop}{Object of class "numeric" or NULL; number of unique markers per cluster used to seed the model, by default 100. If NULL it uses all of them.}
}
\value{
This function returns a list with the initialized matrices H and W.
}
\description{
This functions seeds initialization matrices H and W to perform NMF
}
\examples{

}
