% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{spot_set_opts}
\alias{spot_set_opts}
\alias{spot_get_opts}
\title{Setup}
\usage{
spot_set_opts(
  language = "en",
  response = "list",
  retry = 25,
  sleep = 0.5,
  sleep_if_fail = 3,
  quiet = !interactive()
)

spot_get_opts()
}
\arguments{
\item{language}{Language of text.}

\item{response}{Format of response.}

\item{retry}{Number of times \code{\link{spot_annotate}} should retry the API call if it fails.}

\item{sleep}{Pause between calls in seconds.}

\item{sleep_if_fail}{Pause call attempts in seconds.}

\item{quiet}{Whether to print useful information in console.}
}
\description{
Setup options for session.
}
\section{Languages}{

\itemize{
  \item{\code{en} English}
  \item{\code{de} German}
  \item{\code{nl} Dutch}
  \item{\code{fr} French}
  \item{\code{it} Italian}
  \item{\code{ru} Russian}
  \item{\code{es} Spanish}
  \item{\code{pt} Portuguese}
  \item{\code{hu} Hungarian}
  \item{\code{tr} Turkish}
}
}

\section{Response}{

\itemize{
  \item{\code{list}}
  \item{\code{data.frame}}
}
}

\section{Functions}{

\itemize{
  \item{\code{spot_set_opts} set options}
  \item{\code{spot_get_opts} get otpions}
}
}

\examples{
spot_set_opts(language = "fr")

}
