% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_decon_syn_assessment_fun.R
\name{spatial_decon_syn_assessment_fun}
\alias{spatial_decon_syn_assessment_fun}
\title{This function takes in a seurat object with several tuning parameters and it assesses its performance on synthetic test spots}
\usage{
spatial_decon_syn_assessment_fun(
  se_obj,
  clust_vr,
  verbose = TRUE,
  iter = 2000,
  nstart = 1,
  keep = 100,
  top_dist = 500,
  top_jsd = 5,
  cl_n = 10,
  hvg = 1000,
  al = 0.01,
  ntop = 100,
  n_max = 1e+06,
  k_sub = 8
)
}
\arguments{
\item{se_obj}{Object of class Seurat.}

\item{clust_vr}{Object of class character. Name of the variable containing the cell clustering.}

\item{verbose}{Object of class Logical determining if progress should be reported or not (TRUE by default).}

\item{iter}{Object of class "integer"; number of Gibbs iterations, by default equals 2000.}

\item{nstart}{Object of class "integer". Number of repeated random starts, by default 1.}

\item{keep}{Object of class "integer"; if a positive integer, the log-likelihood is saved every keep iterations, by default 100.}

\item{top_dist}{Object of class integer, on how many top euclidean distance are we going to calculate the JSD, by default 500.}

\item{top_jsd}{Object of class integer, how many of the top spots according JSD distance are we going to use to determine the composition, by default 5.}

\item{cl_n}{Object of class integer, how many cells to grab from each cluster, by default 10.}

\item{hvg}{Object of class integer, how many HVG to pass to the LDA model besides the cluster markers, by default 1000.}

\item{al}{Object of class numeric, which alpha to use to train the model, by default 0.01.}

\item{ntop}{Object of class "numeric"; number of unique markers per cluster used to seed the model, by default 100. If NULL it uses all of them.}

\item{n_max}{Object of class numeric. Max number of synthetic spots to generate, by the default 1*10^6.}

\item{k_sub}{Object of class numeric. Max number of cells per synthetic spot, combination of number of cells per synthetic spot will range between 2:k_sub}
}
\value{
This function returns a list where the first element is the lda model trained, the second is a list with test spot counts + metadata and the third element are the raw_statistics.
}
\description{
This function takes in a seurat object with several tuning parameters and it assesses its performance on synthetic test spots
}
\examples{

}
