% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_decon_syn_assessment_nmf_fun.R
\name{spatial_decon_syn_assessment_nmf_fun}
\alias{spatial_decon_syn_assessment_nmf_fun}
\title{This function takes in a seurat object with several tuning parameters and it assesses its performance on synthetic test spots}
\usage{
spatial_decon_syn_assessment_nmf_fun(
  se_obj,
  clust_vr,
  verbose = TRUE,
  cl_n = 10,
  hvg = 0,
  ntop = 100,
  transf = "uv",
  method = "nsNMF",
  logFC = 1,
  pct1 = 0.9
)
}
\arguments{
\item{se_obj}{Object of class Seurat.}

\item{clust_vr}{Object of class character. Name of the variable containing the cell clustering.}

\item{verbose}{Object of class Logical determining if progress should be reported or not (TRUE by default).}

\item{cl_n}{Object of class integer, how many cells to grab from each cluster, by default 10.}

\item{hvg}{Object of class integer, how many HVG to pass to the LDA model besides the cluster markers, by default 0}

\item{ntop}{Object of class "numeric"; number of unique markers per cluster used to seed the model, by default 100. If NULL it uses all of them.}

\item{transf}{Transformation to normalize the count matrix: cpm (Counts per million), uv (unit variance), sct (Seurat::SCTransform), NULL (no transformation applied). By default CPM.}

\item{method}{Object of class character; Type of method to us to find W and H. Look at NMF package for the options and specifications, by default nsNMF.}

\item{logFC}{Object of class numeric; logFC to filter marker genes by, by default 1.}

\item{pct1}{Object of class numeric; pct1 to filter marker genes by, by default 0.9.}
}
\value{
This function returns a list where the first element is the lda model trained, the second is a list with test spot counts + metadata and the third element are the raw_statistics.
}
\description{
This function takes in a seurat object with several tuning parameters and it assesses its performance on synthetic test spots
}
\examples{

}
