% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{spot}
\alias{spot}
\alias{spot.data.frame}
\alias{spot.character}
\title{Spot}
\usage{
spot(data, text)

\method{spot}{data.frame}(data, text)

\method{spot}{character}(data, text)
}
\arguments{
\item{data}{A data.frame or vector.}

\item{text}{Name of column containing text.}
}
\description{
Spot entities.
}
\examples{
\dontrun{
data <- data.frame(
  text = c(
    "This text is about France.",
    "This text is about Ecology."
  )
)

(results <- spot(data, text))
}

}
