% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{spot_annotate}
\alias{spot_annotate}
\title{Annotate}
\usage{
spot_annotate(
  text,
  url = NULL,
  confidence = NULL,
  support = NULL,
  types = NULL,
  sparql = NULL,
  policy = NULL,
  tidy = TRUE
)
}
\arguments{
\item{text}{A data.frame containing a column named "text" or a vector of class character.}

\item{url}{URL to be annotated.}

\item{confidence}{Confidence score for disambiguation / linking.}

\item{support}{How prominent is this entity in Lucene Model, i.e. number of inlinks in Wikipedia.}

\item{types}{Types filter (Eg.:\code{DBpedia:Place}).}

\item{sparql}{SPARQL filtering.}

\item{policy}{(whitelist) select all entities that have the same type; (blacklist) -
select all entities that have not the same type.}

\item{tidy}{If \code{TRUE} returns a tibble by using \code{\link{tidy_annotations}}.}
}
\description{
Annotate entities.
}
\examples{
\dontrun{
data <- data.frame(
  text = c(
    "This text is about France.",
    "This text is about Ecology."
  )
)

(results <- spot_annotate(data))
}

}
