% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_deconvolution_fun.R
\name{spot_deconvolution}
\alias{spot_deconvolution}
\title{This function processes the data and trains the LDA model}
\usage{
spot_deconvolution(
  lda_mod,
  se_obj,
  clust_vr,
  spot_counts,
  verbose = TRUE,
  ncores = NULL,
  parallelize = TRUE,
  top_dist = 1000,
  top_jsd = 15
)
}
\arguments{
\item{lda_mod}{Object of class LDA_Gibbs.}

\item{se_obj}{Object of class Seurat.}

\item{clust_vr}{Object of class character. Name of the variable containing the cell clustering.}

\item{spot_counts}{Object of class sparse matrix. Count matrix of the spots to deconvolute.}

\item{verbose}{Object of class Logical determining if progress should be reported or not (TRUE by default).}

\item{top_dist}{Object of class integer, on how many top euclidean distance are we going to calculate the JSD.}

\item{top_jsd}{Object of class integer, how many of the top spots according JSD distance are we going to use to determine the composition.}
}
\value{
matrix with the exact predicted composition of each spot.
}
\description{
This function processes the data and trains the LDA model
}
\examples{

}
