% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{spot_df}
\alias{spot_df}
\title{Widen}
\usage{
spot_df(data)
}
\arguments{
\item{data}{Data.frame as returned by \code{\link{spot}}.}
}
\description{
Widen data.frame.
}
\examples{
\dontrun{
results <- spot("This text is about France.")
wide <- spot_df(results)
}


}
