% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{spot_filter}
\alias{spot_filter}
\title{Filter empty documents.}
\usage{
spot_filter(text)
}
\arguments{
\item{text}{vector or list of text.}
}
\description{
Filter empty documents.
}
\examples{
text <- c("Some text", "", "", "More text")

spot_filter(text)

}
