% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{spot_resources}
\alias{spot_resources}
\title{Extract Resources}
\usage{
spot_resources(data)
}
\arguments{
\item{data}{Data as returned by \code{\link{spot_annotate}}.}
}
\description{
Extract Resources from results.
}
\section{Resources extracted}{

\itemize{
  \item{\code{URI}}
  \item{\code{support}}
  \item{\code{types}}
  \item{\code{surfaceFrom}}
  \item{\code{offset}}
  \item{\code{similarityScore}}
  \item{\code{percentageOfSecondRank}}
}
}

\examples{
\dontrun{
results <- spot("This text is about France.")
spot_resources(results)
}

}
