% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{spot_setup}
\alias{spot_setup}
\title{Setup}
\usage{
spot_setup(language = NULL, response = NULL)
}
\arguments{
\item{language}{Language of text.}

\item{response}{Format of response.}
}
\description{
Setup session.
}
\section{Languages}{

\itemize{
  \item{\code{en} English}
  \item{\code{de} German}
  \item{\code{nl} Dutch}
  \item{\code{fr} French}
  \item{\code{it} Italian}
  \item{\code{ru} Russian}
  \item{\code{es} Spanish}
  \item{\code{pt} Portuguese}
  \item{\code{hu} Hungarian}
  \item{\code{tr} Turkish}
}
}

\section{Response}{

\itemize{
  \item{\code{list}}
  \item{\code{data.frame}}
}
}

\examples{
spot_setup(language = "fr")

}
