% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{spot_spot}
\alias{spot_spot}
\alias{spot_spot.data.frame}
\alias{spot_spot.character}
\title{Spot}
\usage{
spot_spot(text, url = NULL, confidence = NULL, support = NULL,
  types = NULL, sparql = NULL, policy = NULL)

\method{spot_spot}{data.frame}(text, url = NULL, confidence = NULL,
  support = NULL, types = NULL, sparql = NULL, policy = NULL)

\method{spot_spot}{character}(text, url = NULL, confidence = NULL,
  support = NULL, types = NULL, sparql = NULL, policy = NULL)
}
\arguments{
\item{text}{A data.frame containing a column named "text" or a vector of class character.}

\item{url}{URL to be annotated.}

\item{confidence}{Confidence score for disambiguation / linking.}

\item{support}{How prominent is this entity in Lucene Model, i.e. number of inlinks in Wikipedia.}

\item{types}{Types filter (Eg.:\code{DBpedia:Place}).}

\item{sparql}{SPARQL filtering.}

\item{policy}{(whitelist) select all entities that have the same type; (blacklist) -
select all entities that have not the same type.}
}
\description{
Spot urls.
}
\examples{
\dontrun{
data <- data.frame(
  text = c(
    "This text is about France.",
    "This text is about Ecology."
  )
)

(results <- spot_spot(data))
}

}
