% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{spot_widen}
\alias{spot_widen}
\title{Widen}
\usage{
spot_widen(data)
}
\arguments{
\item{data}{Data.frame as returned by \code{\link{spot}}.}
}
\description{
Widen data.frame.
}
\examples{
\dontrun{
spot_setup(response = "data.frame")
results <- spot("This text is about France.")
wide <- spot_widen(results)
}


}
