% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotlight_deconvolution.R
\name{spotlight_deconvolution}
\alias{spotlight_deconvolution}
\title{This functions takes in a Seurat object with the scRNAseq data and the count matrix of the spatial transcriptomics data and returns the deconvoluted spots.}
\usage{
spotlight_deconvolution(
  se_sc,
  counts_spatial,
  clust_vr,
  cluster_markers,
  cl_n = 100,
  hvg = 3000,
  ntop = NULL,
  transf = "uv",
  method = "nsNMF",
  min_cont = 0.09
)
}
\arguments{
\item{se_sc}{Object of class Seurat with the scRNAseq data.}

\item{counts_spatial}{Object of class Matrix of shape GENESxSPOT, se_obj\@assays$Spatial@counts.}

\item{clust_vr}{Object of class character. Name of the variable containing the cell clustering.}

\item{cluster_markers}{Object of class dataframe obtained from the function Seurat::FindAllMarkers()}

\item{cl_n}{Object of integer indicating how many cells to keep from each cluster. If a cluster has n < cl_n then all cells will be selected, if it has more then cl_n will be sampled randomly, 100 by default.}

\item{hvg}{Object of class integer, how many HVG to pass to the LDA model besides the cluster markers, by default 0}

\item{ntop}{Object of class "numeric" or NULL; number of unique markers per cluster used to seed the model, by default 100. If NULL it uses all of them.}

\item{transf}{Transformation to normalize the count matrix: cpm (Counts per million), uv (unit variance), sct (Seurat::SCTransform), raw (no transformation applied). By default CPM.}

\item{method}{Object of class character; Type of method to us to find W and H. Look at NMF package for the options and specifications, by default nsNMF.}

\item{min_cont}{Object of class numeric; Indicates the minimum contribution we expect from a cell in that spot. Since we're working with proportions by setting 0.09, by default, means that we will accept those cell types whose weight coefficient is at least 9% of the total.}
}
\value{
This function returns a matrix with the coefficients of the spatial mixtures.
}
\description{
This functions takes in a Seurat object with the scRNAseq data and the count matrix of the spatial transcriptomics data and returns the deconvoluted spots.
}
\examples{

}
