% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_spot_assignment_fun.R
\name{syn_spot_assignment}
\alias{syn_spot_assignment}
\title{This function processes the data and trains the LDA model}
\usage{
syn_spot_assignment(prediction, syn_spots_ls, top_dist = 1000, top_jsd = 15)
}
\arguments{
\item{prediction}{Object of class matrix with the predicted topic probability distributions for each spot. Output from the lda_prediction function.}

\item{syn_spots_ls}{list obtained from the function syn_spot_comb_topic_fun.R. The 1st element is a matrix with topic profiles of all the synthetic spots generated, the 2nd element is the composition of each synthetic spot.}

\item{top_dist}{Object of class integer, on how many top euclidean distance are we going to calculate the JSD.}

\item{top_jsd}{Object of class integer, how many of the top spots according JSD distance are we going to use to determine the composition.}
}
\value{
matrix with the exact predicted composition of each spot.
}
\description{
This function processes the data and trains the LDA model
}
\examples{

}
