% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_spot_comb_topic_fun.R
\name{syn_spot_comb_topic}
\alias{syn_spot_comb_topic}
\title{This function takes in the cluster profiles and returns a sotchastic combination of all the possible ones}
\usage{
syn_spot_comb_topic(
  lda_mod,
  train_cell_clust,
  clust_vr,
  n_max = 1e+06,
  k_sub = 8,
  verbose = TRUE
)
}
\arguments{
\item{lda_mod}{Object of class LDA_Gibbs.}

\item{train_cell_clust}{Object of class vector with cluster of the cells used to train the model.}

\item{clust_vr}{Object of class character. Name of the variable containing the cell clustering.}

\item{n_max}{Object of class numeric. Max number of synthetic spots to generate, by the default 1*10^6.}

\item{k_sub}{Object of class numeric. Max number of cells per synthetic spot, combination of number of cells per synthetic spot will range between 2:k_sub}

\item{verbose}{Object of class Logical determining if progress should be reported or not (TRUE by default).}
}
\value{
This function returns a list where the first element is a matrix with the topic profiles of all possible combinations and the 2nd element is the cell composition of each spot.
}
\description{
This function takes in the cluster profiles and returns a sotchastic combination of all the possible ones
}
\examples{

}
