% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_spot_fun.R
\name{test_spot_fun}
\alias{test_spot_fun}
\title{This functions takes in a seurat object and creates different mixtures resembling spots at different proportions}
\usage{
test_spot_fun(se_obj, clust_vr, n = 10000, verbose = TRUE)
}
\arguments{
\item{se_obj}{seurat object; The seurat object NEEDS to have the variable features calculated}

\item{clust_vr}{Name of the variable containing the cell clustering}

\item{n}{number of spots to generate}

\item{verbose}{Name of the variable containing the cell clustering}
}
\value{
This function returnsa dataframe where each column is the proportion of each spot and a sparse matrix with the synthetic spots with the genenames as rownames and colnames being the spot names
}
\description{
This functions takes in a seurat object and creates different mixtures resembling spots at different proportions
}
\examples{

}
