% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_synthetic_performance_fun.R
\name{test_synthetic_performance}
\alias{test_synthetic_performance}
\title{If you wanna test the performance of your model on synthetic generated test spots you can use this function to benchmark and get a sense of the model's performance.}
\usage{
test_synthetic_performance(test_spots_metadata_mtrx, spot_composition_mtrx)
}
\arguments{
\item{spot_composition_mtrx}{list obtained from the function syn_spot_comb_topic_fun.R. The 1st element is a matrix with topic profiles of all the synthetic spots generated, the 2nd element is the composition of each synthetic spot.}

\item{test_spots_metadata}{Object of class matrix with the predicted topic probability distributions for each spot. Output from the lda_prediction function.}
}
\value{
This function returns a list with TP, TN, FP, FN and the Jensen-Shannon Divergence index.
}
\description{
If you wanna test the performance of your model on synthetic generated test spots you can use this function to benchmark and get a sense of the model's performance.
}
\examples{

}
