% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{tidy_annotations}
\alias{tidy_annotations}
\title{Tidy Results}
\usage{
tidy_annotations(data)
}
\arguments{
\item{data}{Data as returned by \code{\link{spot_annotate}}.}
}
\description{
Tidy results as data.frame.
}
\section{Resources extracted}{

\itemize{
  \item{\code{URI}}
  \item{\code{support}}
  \item{\code{types}}
  \item{\code{surfaceFrom}}
  \item{\code{offset}}
  \item{\code{similarityScore}}
  \item{\code{percentageOfSecondRank}}
}
}

\examples{
\dontrun{
results <- spot("This text is about France.")
tidy_annotations(results)
}

}
