% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_n_predictions_fun.R
\name{top_n_predictions}
\alias{top_n_predictions}
\title{This function gets a distance matrix and returns a list with the indices of the lowest (best) prediction indices}
\usage{
top_n_predictions(dist_mtrx, n)
}
\arguments{
\item{dist_mtrx}{Object of class matrix. Distance matrix to evaluate}

\item{n}{number of best predictions to return}
}
\value{
list of nrows(dist_mtrx) elements with ~n elements each
}
\description{
This function gets a distance matrix and returns a list with the indices of the lowest (best) prediction indices
}
\examples{

}
