% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_profile_per_cluster_nmf_fun.R
\name{topic_profile_per_cluster_nmf}
\alias{topic_profile_per_cluster_nmf}
\title{This function takes in the H coefficient matrix object from and NMF object and returns a matrix object with the topic profile for each cell type}
\usage{
topic_profile_per_cluster_nmf(h, train_cell_clust)
}
\arguments{
\item{h}{Object of class matrix, coefficient matrix from NMF model.}

\item{train_cell_clust}{Object of class vector with cluster of the cells used to train the model.}
}
\value{
This function returns a list where the first element is a matrix with the topic profiles of all possible combinations and the 2nd element is the cell composition of each spot.
}
\description{
This function takes in the H coefficient matrix object from and NMF object and returns a matrix object with the topic profile for each cell type
}
\examples{

}
