% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_viz_fun.R
\name{topic_viz}
\alias{topic_viz}
\title{This function processes the data and trains the LDA model}
\usage{
topic_viz(lda_mod, k, n_terms)
}
\arguments{
\item{lda_mod}{Object of class LDA_Gibbs.}

\item{k}{topic we want to visualize.}

\item{n_terms}{number of terms to show.}
}
\value{
The n_terms most likely genes defining that cluster.
}
\description{
This function processes the data and trains the LDA model
}
\examples{

}
