% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_nmf.R
\name{train_nmf}
\alias{train_nmf}
\title{This functions carries out the NMF and returns and NMF object.}
\usage{
train_nmf(
  cluster_markers,
  se_sc,
  mtrx_spatial,
  clust_vr,
  ntop = NULL,
  transf = "cpm",
  method = "nsNMF",
  hvg = 3000,
  assay = "RNA",
  slot = "counts"
)
}
\arguments{
\item{cluster_markers}{Object of class dataframe obtained from the function Seurat::FindAllMarkers()}

\item{se_sc}{Object of class Seurat with the scRNAseq data.}

\item{mtrx_spatial}{Object of class Matrix of shape GENESxSPOT.}

\item{clust_vr}{Object of class character; Name of the variable containing the cell clustering.}

\item{ntop}{Object of class "numeric" or NULL; number of unique markers per cluster used to seed the model, by default 100. If NULL it uses all of them.}

\item{transf}{Transformation to normalize the count matrix: cpm (Counts per million), uv (unit variance), raw (no transformation applied). By default CPM.}

\item{method}{Object of class character; Type of method to us to find W and H. Look at NMF package for the options and specifications, by default nsNMF.}

\item{hvg}{Object of class numeric or "uns"; Number of highly variable genes to use on top of the marker genes, if "uns" then it is completely unsupervised and use top 3000 HVG.}

\item{assay}{Object of class character; From which assay to grab the expression data to train the model, by default "RNA".}

\item{slot}{Object of class character; From which slot to grab the expression data to train the model, by default "counts".}
}
\value{
This function returns a list with the initialized matrices H and W.
}
\description{
This functions carries out the NMF and returns and NMF object.
}
\examples{

}
