% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_se_obj_fun.R
\name{downsample_se_obj}
\alias{downsample_se_obj}
\title{This function downsamples the number of cells and genes used to train the model}
\usage{
downsample_se_obj(se_obj, clust_vr, cluster_markers_all, cl_n = 100)
}
\arguments{
\item{se_obj}{Object of class Seurat with the data of interest}

\item{clust_vr}{Name of the variable containing the cell clustering}

\item{cluster_markers_all}{Object of class dataframe obtained from the function Seurat::FindAllMarkers()}

\item{cl_n}{Object of integer indicating how many cells to keep from each cluster. If a cluster has n < cl_n then all cells will be selected, if it has more then cl_n will be sampled randomly, 100 by default.}
}
\value{
A downsampled Seurat object from the original
}
\description{
This function downsamples the number of cells and genes used to train the model
}
