% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotImage.R
\name{plotImage}
\alias{plotImage}
\alias{plotImage,character-method}
\alias{plotImage,Seurat-method}
\alias{plotImage,SpatialExperiment-method}
\alias{plotImage,array-method}
\alias{plotImage,ANY-method}
\title{Plot JP(E)G/PNG/Raster/RGB images}
\usage{
\S4method{plotImage}{character}(x)

\S4method{plotImage}{Seurat}(x, ..., slice = Images(x)[1])

\S4method{plotImage}{SpatialExperiment}(x, ..., slice = imgData(x)[1, "sample_id"])

\S4method{plotImage}{array}(x)

\S4method{plotImage}{ANY}(x)
}
\arguments{
\item{x}{A variety of objects can be passed: character string corresponding
to an image file path, valid file types are JPG, JPEG and PNG. It can also
take as input objects of class raster and RGB arrays. It can also take
a SpatialExperiment or Seurat object from which the image will be extracted.}

\item{slice}{Character string indicating which image slice to use when
SpatialExperiment or Seurat objects are passed. By default uses the first
slice available.}

\item{alpha}{NOT IMPLEMENTED - single numeric between 0 and 1 determining the
image opacity. Lower values correspond to more transparency.}
}
\value{
\code{ggplot} object
}
\description{
This function takes in an image-related object - path to
  JP(E)G/PNG file, raster object, RGBarray. It returns a ggplot object with
  the selected image.
}
\examples{
# Filename
plotImage("../../inst/extdata/SPOTlight.png")
path <- paste0(system.file(package = "SPOTlight"), "/spatial/tissue_lowres_image.png")
plotImage(path)
# array
png_img <- png::readPNG(path)
plotImage(png_img)
# Seurat Object
so <- LoadData("stxBrain", type = "anterior1")
plotImage(so)
# SpatialExperiment
spe0 <- as.SingleCellExperiment(LoadData("stxBrain", type = "anterior1"))
plotImage(spe0)
}
\author{
Marc Elosua Bayes & Helena L Crowell
}
