% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpatialscatterpie.R
\name{plotSpatialScatterpie}
\alias{plotSpatialScatterpie}
\alias{plotSpatialScatterpie,matrix,matrix-method}
\alias{plotSpatialScatterpie,Seurat,ANY-method}
\alias{plotSpatialScatterpie,SpatialExperiment,ANY-method}
\alias{plotSpatialScatterpie,data.frame,ANY-method}
\alias{plotSpatialScatterpie,ANY,data.frame-method}
\alias{plotSpatialScatterpie,ANY,ANY-method}
\title{Spatial scatterpie}
\usage{
\S4method{plotSpatialScatterpie}{matrix,matrix}(
  x,
  y,
  ...,
  cell_types = colnames(y),
  img = FALSE,
  scatterpie_alpha = 1,
  pie_scale = 0.4
)

\S4method{plotSpatialScatterpie}{Seurat,ANY}(x, y, ..., slice = Images(x)[1], img = FALSE)

\S4method{plotSpatialScatterpie}{SpatialExperiment,ANY}(
  x,
  y,
  ...,
  slice = imgData(x)[1, "sample_id"],
  img = FALSE
)

\S4method{plotSpatialScatterpie}{data.frame,ANY}(x, y, ...)

\S4method{plotSpatialScatterpie}{ANY,data.frame}(x, y, ...)

\S4method{plotSpatialScatterpie}{ANY,ANY}(x)
}
\arguments{
\item{x}{Object containig the spots coordinates, it can be an object of class
SpatialExperiment, Seurat, dataframe or matrix. For the latter two
rownames should have the spot barcodes to match x.}

\item{y}{Matrix or dataframe containing the deconvoluted spots. rownames
hould have the spot barcodes to match x.}

\item{cell_types}{Vector of cell type names to plot. By default uses the
column names of y.}

\item{img}{Logical TRUE or FALSE indicating whether to plot the image or not.
Objects of classes accepted by \code{plotImage} can also be passed and
that image will be used. By default FALSE.}

\item{scatterpie_alpha}{Numeric scalar to set the alpha of the pie charts.
By default 1.}

\item{slice}{Character string indicating which slice to plot if img is TRUE.
By default uses the first image.}

\item{pie_scaleNumeric}{scalar to set the size of the pie charts.
By default 0.4.}
}
\value{
\code{ggplot} object
}
\description{
This function takes in the coordinates of the spots and the
  proportions of the cell types within each spot. It returns a plot with
  each spot as a piechart of the proportions of the cell type composition.
}
\examples{
# Filename
plotSpatialScatterpie("~/packages/SPOTlight/inst/extdata/SPOTlight.png")
path <- paste0(system.file(package="SPOTlight"), "/spotlight_ls_anterior.RDS")
y <- readRDS(path)[[2]]

# array
# Seurat Object
x <- LoadData("stxBrain", type = "anterior1")
plotSpatialScatterpie(x, y)
# SpatialExperiment
example(read10xVisium, echo = FALSE)
plotSpatialScatterpie(spe, y)
}
