% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDeconvolution.R
\name{runDeconvolution}
\alias{runDeconvolution}
\title{Run Deconvolution using NNLS model}
\usage{
runDeconvolution(
  x,
  mod,
  ref,
  scale = TRUE,
  min_prop = 0.01,
  verbose = TRUE,
  assay = "RNA",
  slot = "counts",
  L1 = 0.1,
  L2 = 0
)
}
\arguments{
\item{x}{mixture dataset. Can be a numeric matrix,
\code{SingleCellExperiment}, \code{SpatialExperiment} or 
\code{SeuratObjecy}.}

\item{mod}{object of class NMFfit as obtained from trainNMF.}

\item{ref}{bject of class matrix containing the topic profiles for each cell
type as obtained from trainNMF.}

\item{scale}{logical specifying whether to scale single-cell counts to unit
variance. This gives the user the option to normalize the data beforehand
as you see fit (CPM, FPKM, ...) when passing a matrix or specifying the
slot from where to extract the count data.}

\item{min_prop}{scalar in [0,1] setting the minimum contribution
expected from a cell type in \code{x} to observations in \code{y}.
By default 0.}

\item{verbose}{logical. Should information on progress be reported?}

\item{assay}{if the object is of Class \code{Seurat}, character string
specifying the assay from which to extract the expression matrix.
By default "RNA". Ignore for the rest of x input classes.}

\item{slot}{if the object is of Class \code{Seurat}, character string
specifying the slot from which to extract the expression matrix. If the
object is of class \code{SpatialExperiment} indicates matrix to use.
By default "counts".}

\item{L1}{L1/LASSO penalty to be subtracted from b. See ?RcppML::nnls()
for more info.}

\item{L2}{Ridge penalty to be added to diagonal of a. See ?RcppML::nmf()
for more info.}
}
\value{
base a list where the first element is an \code{NMFfit} object and
  the second is a matrix contatining the topic profiles learnt.
}
\description{
This function takes in the mixture data, the trained model & the
  topic profiles and returns the proportion of each cell type within each
   mixture
}
\examples{
set.seed(321)
# mock up some single-cell, mixture & marker data
sce <- mockSC(ng = 200, nc = 10, nt = 3)
spe <- mockSP(sce)
mgs <- getMGS(sce)

res <- trainNMF(
    x = sce,
    y = spe,
    groups = sce$type,
    mgs = mgs,
    weight_id = "weight",
    group_id = "type",
    gene_id = "gene")
# Run deconvolution
decon <- runDeconvolution(
    x = spe,
    mod = res[["mod"]],
    ref = res[["topic"]])
}
\author{
Marc Elosua Bayes & Helena L Crowell
}
