% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_decon_syn_assessment_fun.R
\name{spatial_decon_syn_assessment_fun}
\alias{spatial_decon_syn_assessment_fun}
\title{This function takes in a seurat object with several tuning parameters and it assesses its performance on synthetic test spots}
\usage{
spatial_decon_syn_assessment_fun(
  se_obj,
  clust_vr,
  verbose = TRUE,
  iter = 3000,
  nstart = 1,
  keep = 100,
  top_dist = 1000,
  top_jsd = 15,
  cl_n = 100
)
}
\arguments{
\item{se_obj}{Object of class Seurat.}

\item{clust_vr}{Object of class character. Name of the variable containing the cell clustering.}

\item{verbose}{Object of class Logical determining if progress should be reported or not (TRUE by default).}

\item{iter}{Object of class "integer"; number of Gibbs iterations, by default equals 5000.}

\item{nstart}{Object of class "integer". Number of repeated random starts.}

\item{keep}{Object of class "integer"; if a positive integer, the log-likelihood is saved every keep iterations, by default 100.}

\item{top_dist}{Object of class integer, on how many top euclidean distance are we going to calculate the JSD.}

\item{top_jsd}{Object of class integer, how many of the top spots according JSD distance are we going to use to determine the composition.}

\item{cl_n}{Object of class integer, how many cells to grab from each cluster.}
}
\value{
This function returns a list where the first element is the lda model trained, the second is a list with test spot counts + metadata and the third element are the raw_statistics.
}
\description{
This function takes in a seurat object with several tuning parameters and it assesses its performance on synthetic test spots
}
\examples{

}
