% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_decon_syn_assessment_nmf_fun.R
\name{spatial_decon_syn_assessment_nmf_fun}
\alias{spatial_decon_syn_assessment_nmf_fun}
\title{This function takes in a seurat object with several tuning parameters and it assesses its performance on synthetic test spots}
\usage{
spatial_decon_syn_assessment_nmf_fun(
  se_sc,
  clust_vr,
  cluster_markers,
  n_syn_mixt = 1000,
  cl_n = 100,
  hvg = 3000,
  ntop = NULL,
  transf = "uv",
  method = "nsNMF",
  min_cont = 0.01,
  assay = "RNA",
  slot = "counts"
)
}
\arguments{
\item{se_sc}{Object of class Seurat with the scRNAseq data.}

\item{clust_vr}{Object of class character. Name of the variable containing the cell clustering.}

\item{cluster_markers}{Object of class dataframe obtained from the function Seurat::FindAllMarkers().}

\item{n_syn_mixt}{Object of class integer specifying how many synthetic mixtures to generate.}

\item{cl_n}{Object of integer indicating how many cells to keep from each cluster. If a cluster has n < cl_n then all cells will be selected, if it has more then cl_n will be sampled randomly, 100 by default.}

\item{hvg}{Object of class numeric or "uns"; Number of highly variable genes to use on top of the marker genes, if "uns" then it is completely unsupervised and use top 3000 HVG.}

\item{ntop}{Object of class "numeric" or NULL; number of unique markers per cluster used to seed the model, by default NULL. If NULL it uses all of them.}

\item{transf}{Transformation to normalize the count matrix: cpm (Counts per million), uv (unit variance), sct (Seurat::SCTransform), raw (no transformation applied). By default CPM.}

\item{method}{Object of class character; Type of method to us to find W and H. Look at NMF package for the options and specifications, by default nsNMF.}

\item{min_cont}{Object of class numeric; Indicates the minimum contribution we expect from a cell in that spot. Since we're working with proportions by setting 0.01, by default, means that we will accept those cell types whose weight coefficient is at least 1% of the total.}
}
\value{
This function returns a list where the first element is a list with the NMF model trained and the cell labels, the second is a list with the raw_statistics.
}
\description{
This function takes in a seurat object with several tuning parameters and it assesses its performance on synthetic test spots
}
\examples{

}
