% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_deconvolution_fun.R
\name{spot_deconvolution}
\alias{spot_deconvolution}
\title{This function processes the data and trains the LDA model}
\usage{
spot_deconvolution(
  lda_mod,
  se_obj,
  clust_vr,
  verbose,
  prediction,
  syn_spots_ls,
  top_dist = 1000,
  top_JSD = 15
)
}
\arguments{
\item{lda_mod}{Object of class LDA_Gibbs.}

\item{se_obj}{Object of class Seurat.}

\item{clust_vr}{Object of class character. Name of the variable containing the cell clustering.}

\item{verbose}{Object of class Logical determining if progress should be reported or not (TRUE by default).}

\item{prediction}{Object of class matrix with the predicted topic probability distributions for each spot. Output from the lda_prediction function.}

\item{syn_spots_ls}{list obtained from the function syn_spot_comb_topic_fun.R. The 1st element is a matrix with topic profiles of all the synthetic spots generated, the 2nd element is the composition of each synthetic spot.}

\item{top_dist}{Object of class integer, on how many top euclidean distance are we going to calculate the JSD.}

\item{top_JSD}{Object of class integer, how many of the top spots according JSD distance are we going to use to determine the composition.}
}
\value{
matrix with the exact predicted composition of each spot.
}
\description{
This function processes the data and trains the LDA model
}
\examples{

}
