% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_deconvolution_fun.R
\name{spot_deconvolution}
\alias{spot_deconvolution}
\title{This function processes the data and trains the LDA model}
\usage{
spot_deconvolution(
  lda_mod,
  train_cell_clust,
  clust_vr,
  spot_counts,
  verbose = TRUE,
  ncores = NULL,
  top_dist = 1000,
  top_jsd = 10,
  n_max = 1e+06,
  k_sub = 8
)
}
\arguments{
\item{lda_mod}{Object of class LDA_Gibbs.}

\item{train_cell_clust}{Object of class vector with cluster of the cells used to train the model.}

\item{clust_vr}{Object of class character. Name of the variable containing the cell clustering.}

\item{spot_counts}{Object of class sparse matrix. Count matrix of the spots to deconvolute.}

\item{verbose}{Object of class Logical determining if progress should be reported or not (TRUE by default).}

\item{ncores}{Object of class integer, How many cores to use to parallelize the process, if NULL i the process will be run sequentially.}

\item{top_dist}{Object of class integer, on how many top euclidean distance are we going to calculate the JSD.}

\item{top_jsd}{Object of class integer, how many of the top spots according JSD distance are we going to use to determine the composition.}

\item{n_max}{Object of class numeric. Max number of synthetic spots to generate, by the default 1*10^6.}

\item{k_sub}{Object of class numeric. Max number of cells per synthetic spot, combination of number of cells per synthetic spot will range between 2:k_sub}
}
\value{
matrix with the exact predicted composition of each spot.
}
\description{
This function processes the data and trains the LDA model
}
\examples{

}
