% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPOTlight.R
\name{SPOTlight}
\alias{SPOTlight}
\alias{SPOTlight,SingleCellExperiment,ANY-method}
\alias{SPOTlight,ANY,SingleCellExperiment-method}
\alias{SPOTlight,Seurat,ANY-method}
\alias{SPOTlight,ANY,Seurat-method}
\alias{SPOTlight,matrix,matrix-method}
\title{Deconvolution of mixture using single-cell data}
\usage{
\S4method{SPOTlight}{SingleCellExperiment,ANY}(
  x,
  y,
  ...,
  assay = "counts",
  groups = colLabels(x, onAbsence = "error")
)

\S4method{SPOTlight}{ANY,SingleCellExperiment}(x, y, ..., assay = "counts")

\S4method{SPOTlight}{Seurat,ANY}(x, y, ..., slot = "counts", assay = "RNA", groups = Idents(x))

\S4method{SPOTlight}{ANY,Seurat}(x, y, ..., slot = "counts", assay = "RNA")

\S4method{SPOTlight}{matrix,matrix}(
  x,
  y,
  groups,
  mgs,
  n_top = NULL,
  gene_id = "gene",
  group_id = "cluster",
  weight_id = "weight",
  scale = TRUE,
  model = c("ns", "std"),
  min_prop = 0.01,
  verbose = TRUE
)
}
\arguments{
\item{x, y}{single-cell and mixture dataset, respectively. Can be a 
numeric matrix, \code{SingleCellExperiment} or \code{SeuratObjecy}.}

\item{groups}{vector of group labels for cells in \code{x}. 
When \code{x} is a \code{SingleCellExperiment} or \code{SeuratObject}, 
defaults to \code{colLabels} and \code{Idents(x)}, respectively.}

\item{mgs}{\code{data.frame} or \code{DataFrame} of marker genes. 
Must contain columns holding gene identifiers, group labels and
the weight (e.g., logFC, -log(p-value) a feature has in a given group.}

\item{gene_id, group_id, weight_id}{character specifying the column in \code{mgs}
containing gene identifiers, group labels and weights, respectively.}

\item{scale}{logical specifying whether to
scale single-cell counts to unit variance.}

\item{min_prop}{scalar in [0,1] setting the minimum contribution
expected from a cell type in \code{x} to observations in \code{y}.}

\item{verbose}{logical. Should information on progress be reported?}
}
\description{
...
}
\examples{
print("todo")
}
\author{
Marc Elosua Bayes
}
