% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPOTlight.R
\name{SPOTlight}
\alias{SPOTlight}
\alias{SPOTlight,SingleCellExperiment,ANY-method}
\alias{SPOTlight,ANY,SingleCellExperiment-method}
\alias{SPOTlight,Seurat,ANY-method}
\alias{SPOTlight,ANY,Seurat-method}
\alias{SPOTlight,ANY,dgCMatrix-method}
\alias{SPOTlight,dgCMatrix,ANY-method}
\alias{SPOTlight,matrix,matrix-method}
\title{Deconvolution of mixture using single-cell data}
\usage{
\S4method{SPOTlight}{SingleCellExperiment,ANY}(
  x,
  y,
  ...,
  assay = "counts",
  groups = colLabels(x, onAbsence = "error")
)

\S4method{SPOTlight}{ANY,SingleCellExperiment}(x, y, ..., assay = "counts")

\S4method{SPOTlight}{Seurat,ANY}(x, y, ..., slot = "counts", assay = "RNA", groups = Idents(x))

\S4method{SPOTlight}{ANY,Seurat}(x, y, ..., slot = "counts", assay = "RNA")

\S4method{SPOTlight}{ANY,dgCMatrix}(x, y, ..., slot = "counts", assay = "RNA")

\S4method{SPOTlight}{dgCMatrix,ANY}(x, y, ..., slot = "counts", assay = "RNA")

\S4method{SPOTlight}{matrix,matrix}(
  x,
  y,
  groups,
  mgs,
  n_top = NULL,
  gene_id = "gene",
  group_id = "cluster",
  weight_id = "weight",
  hvg = NULL,
  scale = TRUE,
  model = c("ns", "std"),
  min_prop = 0.01,
  verbose = TRUE
)
}
\arguments{
\item{x, y}{single-cell and mixture dataset, respectively. Can be a 
numeric matrix, \code{SingleCellExperiment} or \code{SeuratObjecy}.}

\item{groups}{vector of group labels for cells in \code{x}. 
When \code{x} is a \code{SingleCellExperiment} or \code{SeuratObject}, 
defaults to \code{colLabels} and \code{Idents(x)}, respectively.}

\item{mgs}{\code{data.frame} or \code{DataFrame} of marker genes. 
Must contain columns holding gene identifiers, group labels and
the weight (e.g., logFC, -log(p-value) a feature has in a given group.}

\item{gene_id, group_id, weight_id}{character specifying the column in \code{mgs}
containing gene identifiers, group labels and weights, respectively.}

\item{hvg}{character vector containing hvg to include in the model. 
By default NULL.}

\item{scale}{logical specifying whether to scale single-cell counts to unit
variance. This gives the user the option to normalize the data beforehand
as you see fit (CPM, FPKM, ...) when passing a matrix or specifying the
slot from where to extract the count data.}

\item{min_prop}{scalar in [0,1] setting the minimum contribution
expected from a cell type in \code{x} to observations in \code{y}.
By default 0.}

\item{verbose}{logical. Should information on progress be reported?}
}
\value{
a numeric matrix with rows corresponding to samples and columns to groups
}
\description{
This is the backbone function which takes in single cell
  expression data to deconvolute spatial transcriptomics spots.
}
\details{
SPOTlight uses a Non-Negative Matrix Factorization approach to learn
  which genes are important for each cell type. In order to drive the
  factorization and give more importance to cell type marker genes we
  previously compute them and use them to initialize the basis matrix. This
  initialized matrices will then be used to carry out the factorization with 
  the single cell expression data. Once the model has learn the topic
  profiles for each cell type we use non-negative least squares (NNLS) to
  obtain the topic contributions to each spot. Lastly, NNLS is again used to 
  obtain the proportion of each cell type for each spot by finding the
  fitting the single-cell topic profiles to the spots topic contributions.
}
\examples{
# Get sc Brain data from 
# TENxBrainData
# http://127.0.0.1:25293/library/TENxBrainData/doc/TENxBrainData.html
# ExperimentHub
# https://bioconductor.org/packages/release/bioc/vignettes/ExperimentHub/inst/doc/ExperimentHub.html#using-experimenthub-to-retrieve-data
library(TENxBrainData)
library(ExperimentHub)
sc <- TENxBrainData20k()

# Get Visium data from TENxVisiumData
# https://bioconductor.org/packages/release/data/experiment/vignettes/TENxVisiumData/inst/doc/vignette.html
library(TENxVisiumData)
TODO 
print("todo")
}
\author{
Marc Elosua-Bayes & Helena L. Crowell
}
