% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_spot_comb_topic_fun.R
\name{syn_spot_comb_topic}
\alias{syn_spot_comb_topic}
\title{This function takes in the cluster profiles and returns a sotchastic combination of all the possible ones}
\usage{
syn_spot_comb_topic(lda_mod, se_obj, clust_vr, verbose = TRUE)
}
\arguments{
\item{lda_mod}{Object of class LDA_Gibbs.}

\item{se_obj}{Object of class Seurat.}

\item{clust_vr}{Object of class character. Name of the variable containing the cell clustering.}

\item{verbose}{Object of class Logical determining if progress should be reported or not (TRUE by default).}
}
\value{
This function returns a list where the first element is a matric with the topic profiles of all possible combinations and the 2nd element is the cell composition of each spot.
}
\description{
This function takes in the cluster profiles and returns a sotchastic combination of all the possible ones
}
\examples{

}
