% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_synthetic_performance_fun.R
\name{test_synthetic_performance}
\alias{test_synthetic_performance}
\title{If you wanna test the performance of your model on synthetic generated test spots you can use this function to benchmark and get a sense of the model's performance.}
\usage{
test_synthetic_performance(test_spots_metadata_mtrx, spot_composition_mtrx)
}
\arguments{
\item{spot_composition_mtrx}{Object of class matrix with the predicted topic probability distributions for each spot.}

\item{test_spots_metadata}{Object of class matrix containing the ground truth composition of each spot as obtained from the function syn_spot_comb_topic_fun.R, 2nd element.}
}
\value{
This function returns a list with TP, TN, FP, FN and the Jensen-Shannon Divergence index.
}
\description{
If you wanna test the performance of your model on synthetic generated test spots you can use this function to benchmark and get a sense of the model's performance.
}
\examples{

}
