% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_LDA_fun.R
\name{train_lda}
\alias{train_lda}
\title{This function processes the data and trains the LDA model}
\usage{
train_lda(
  se_obj,
  clust_vr,
  cluster_markers,
  al = 0.01,
  ntop = 100,
  verbose = 1,
  estimate.beta = TRUE,
  save = 0,
  keep = 100,
  nstart = 1,
  best = TRUE,
  delta = 0.1,
  iter = 2000,
  burnin = 0,
  thin = NULL,
  ...
)
}
\arguments{
\item{se_obj}{Object of class Seurat with the data of interest}

\item{clust_vr}{Name of the variable containing the cell clustering}

\item{cluster_markers}{Object of class dataframe obtained from the function Seurat::FindAllMarkers()}

\item{ntop}{Object of class "numeric"; number of unique markers per cluster used to seed the model, by default 100. If NULL it uses all of them.}

\item{verbose}{Object  of  class "integer".   If  a  positive  integer,  then  the  progress  is  reported  every "integer" verbose iterations. If 0 (default), no output is generated during model fitting}

\item{estimate.beta}{Object of class "logical"; controls if beta, the term distribution of the topics, isfixed, by default equals TRUE.}

\item{save}{Object of class "integer".  If a positive integer the estimated model is saved all verbose iterations. If 0 (default), no output is generated during model fitting.}

\item{keep}{Object of class "integer"; if a positive integer, the log-likelihood is saved every keep iterations, by default 100.}

\item{nstart}{Object of class "integer". Number of repeated random starts.}

\item{best}{Object of class "logical"; if TRUE only the model with the maximum (posterior) likelihood is returned, by default equals TRUE.}

\item{delta}{Object of class "numeric"; initial value for delta, by default equals 0.1.}

\item{iter}{Object of class "integer"; number of Gibbs iterations, by default equals 5000.}

\item{burnin}{Object of class "integer"; number of omitted Gibbs iterations at beginning, by default equals 0.}

\item{thin}{Object of class "integer"; number of omitted in-between Gibbs iterations, by default equals iter.}

\item{...}{Other arguments passed on to LDA mod}

\item{alpha}{Object of class "numeric"; initial value for alpha, by default 0.01.}
}
\value{
A Latent Dirichlet Allocation list of model/s depending on if return best is T or F + a vector of the clusters of the cells used to train the model.
}
\description{
This function processes the data and trains the LDA model
}
\examples{

}
